requirements.txt
accelerate==1.1.1
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
openai==1.54.4
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.3.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.0.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
